Rename `CodePreconditions` to `Preconditions2` and use it instead of `Validate`
Also introduced constant for field name type separator.
