+++ base/src/main/java/io/spine/type/TypeUrl.java
@@ -40,7 +40,7 @@ import java.util.Objects;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.protobuf.Internal.getDefaultInstance;
-import static io.spine.validate.Validate.checkNotEmptyOrBlank;
+import static io.spine.util.Preconditions2.checkNotEmptyOrBlank;
 import static java.lang.String.format;
 
 /**
@@ -68,8 +68,8 @@ public final class TypeUrl implements Serializable {
     private final String typeName;
 
     private TypeUrl(String prefix, String typeName) {
-        this.prefix = checkNotEmptyOrBlank(prefix, "typeUrlPrefix");
-        this.typeName = checkNotEmptyOrBlank(typeName, "typeName");
+        this.prefix = checkNotEmptyOrBlank(prefix);
+        this.typeName = checkNotEmptyOrBlank(typeName);
     }
 
     /**
